IGNITE-24261 Extract name utils
